backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-genai>=0.2.2
sentence-transformers>=2.7.0
httpx>=0.27.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
pydantic>=2.7.0
markdown>=3.6
psycopg2-binary>=2.9.9
pgvector>=0.2.5
